+++ src/main/java/mod/gottsch/forge/magic_treasures/core/integration/MagicTreasuresIntegrations.java
@@ -36,7 +36,7 @@ public class MagicTreasuresIntegrations {
 
 	public static void registerCuriosIntegration() {
 		IEquipmentSpellHandler equipmentSpellHandler = null;
-		if (Config.SERVER.integration.enableCurios.get() && ModList.get().isLoaded("curios")) {
+		if (ModList.get().isLoaded("curios")) {
 			MagicTreasures.LOGGER.debug("curios IS loaded");
 			try {
 				equipmentSpellHandler =
@@ -49,11 +49,14 @@ public class MagicTreasuresIntegrations {
 			MagicTreasures.LOGGER.debug("equipmentHandler is null");
 			equipmentSpellHandler = new HotbarEquipmentSpellHandler();
 		}
+		// TODO might have to register this earlier and set the spell handler here
 		MinecraftForge.EVENT_BUS.register(new SpellEventHandler(equipmentSpellHandler));
 	}
 
 	public static void registerTreasure2Integration() {
 		if (ModList.get().isLoaded("treasure2")) {
+			MagicTreasures.LOGGER.debug("treasure2 IS loaded");
+			MagicTreasures.LOGGER.debug("registering t2 injectable loot tables");
 			TreasureApi.registerLootTables(MagicTreasures.MOD_ID);
 		}
 	}